加盟商功能

+++ JCSJGL/Page_Cangku.aspx.cs
@@ -9,8 +9,13 @@ using System.Web.UI.WebControls;
 
 namespace JCSJGL
 {
-    public partial class Page_Cangku : System.Web.UI.Page
+    public partial class Page_Cangku : MyPage
     {
+        public Page_Cangku()
+        {
+            _PageName = PageName.仓库信息;
+        }
+
         protected void Page_Load(object sender, EventArgs e)
         {
             //初始化
@@ -40,28 +45,50 @@ namespace JCSJGL
         /// <param name="id"></param>
         private void deleteCangku(int id)
         {
+            //检查权限
+            Authenticate.CheckOperation(_PageName, PageOpt.删除, _LoginUser);
+
             DBContext db = new DBContext();
+            TCangku oc = db.GetCangkuById(id);
+            if (oc.jmsid != _LoginUser.jmsid)
+            {
+                throw new MyException("非法操作，无法删除此仓库");
+            }
             db.DeleteCangku(id);
 
             loadCangkus();
         }
 
         /// <summary>
-        /// 加载所有仓库信息
+        /// 加载仓库信息
         /// </summary>
         private void loadCangkus()
         {
             DBContext db = new DBContext();
-            TCangku[] cs = db.GetCangkus();
+            TCangku[] cs =null;
+            if (_LoginUser.juese == (byte)Tool.JCSJ.DBCONSTS.USER_XTJS.系统管理员)
+            {
+                cs = db.GetCangkus();
+                grid_cangku.Columns.Insert(1, new BoundField()
+                {
+                    HeaderText = "加盟商",
+                    DataField = "jiamengshang"
+                });
+            }
+            else
+            {
+                cs = db.GetCangkusByJmcId(_LoginUser.jmsid);
+            }
+
             var dfs = cs.Select(r => new
             {
                 id = r.id,
+                jiamengshang = r.TJiamengshang.mingcheng,
                 mingcheng = r.mingcheng,
                 dizhi = r.dizhi,
                 lianxiren = r.lianxiren,
                 dianhua = r.dianhua,
                 beizhu = r.beizhu,
-                caozuoren = r.TUser.yonghuming,
                 charushijian = r.charushijian,
                 xiugaishijian = r.xiugaishijian,
                 editParams = r.id + ",'" + r.mingcheng + "','" + r.dizhi + "','" + r.lianxiren + "','" + r.dianhua + "','" + r.beizhu + "'"
@@ -79,12 +106,19 @@ namespace JCSJGL
         /// <param name="e"></param>
         protected void btn_edit_Click(object sender, EventArgs e)
         {
+            Authenticate.CheckOperation(_PageName, PageOpt.修改, _LoginUser);
+
             TCangku f = getEditInfo();
             f.id = int.Parse(hid_id.Value);
-            f.caozuorenid = ((TUser)Session["USER"]).id;
+            f.caozuorenid = _LoginUser.id;
             f.xiugaishijian = DateTime.Now;
 
             DBContext db = new DBContext();
+            TCangku oc = db.GetCangkuById(f.id);
+            if (oc.jmsid != _LoginUser.jmsid)
+            {
+                throw new MyException("非法操作，无法修改该仓库的信息");
+            }
             db.UpdateCangku(f);
 
             loadCangkus();
@@ -121,12 +155,21 @@ namespace JCSJGL
         /// <param name="e"></param>
         protected void btn_add_Click(object sender, EventArgs e)
         {
+            Authenticate.CheckOperation(_PageName, PageOpt.增加, _LoginUser);
+
             TCangku f = getEditInfo();
-            f.caozuorenid = ((TUser)Session["USER"]).id;
+            f.jmsid = _LoginUser.jmsid;
+            f.caozuorenid = _LoginUser.id;
             f.charushijian = DateTime.Now;
             f.xiugaishijian = DateTime.Now;
 
             DBContext db = new DBContext();
+            //限制可增加的仓库数量
+            TCangku[] cs = db.GetCangkusByJmcId(_LoginUser.jmsid);
+            if (cs.Count() >= _LoginUser.TJiamengshang.cangkushu)
+            {
+                throw new MyException("拥有的仓库数量已到上限，如有需要增加更多仓库请联系系统管理员");
+            }
             db.InsertCangku(f);
 
             loadCangkus();
